Refactor determiner service and restructure grammar constants
In this commit, the DeterminerServiceService was renamed to DeterminerService for naming consistency. Also, all grammar constants (Gender, Case, Number, etc.) were moved from a single file 'GermanGrammarConstants' to separate files inside 'grammar-constants' directory to improve code organization.

Additionally, a new functionality was added in DeterminerService, which now holds a list of German definite articles initialized on creation and a function to filter these articles based on provided criteria.

The refactor was done to improve understandability and maintainability of the code including testing. The new function eases filtering of German articles, accelerating the language processing tasks.